geminiplayground/pyproject.toml
[tool.poetry]
name = "geminiplayground"
version = "0.10.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"
packages = [{ include = "geminiplayground", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
google-api-python-client = "^2.123.0"
pillow = "^10.3.0"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"
matplotlib = "^3.8.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
tqdm = "^4.66.2"
pydantic = "^2.6.4"
rich = "^13.7.1"
alive-progress = "^3.1.5"
pygithub = "^2.3.0"
gitpython = "^3.1.42"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
typer = "^0.12.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.29"
jinja2 = "^3.1.3"
more-itertools = "^10.2.0"
markdown = "^3.6"
pygments = "^2.17.2"
vertexai = "^1.46.0"
requests-toolbelt = "^1.0.0"
joblib = "^1.3.2"
diskcache = "^5.6.3"
pytest = "^8.1.1"
pydantic-settings = "^2.2.1"
greenlet = "^3.0.3"
gunicorn = "^22.0.0"
pypdf2 = "^3.0.1"
pymupdf = "^1.24.4"
google-generativeai = "^0.5.4"
tenacity = "^8.3.0"
pandas = "^2.2.2"
validators = "0.28.1"
langchain-google-genai = "^1.0.5"


[tool.poetry.group.rag]
optional = true


[tool.poetry.group.rag.dependencies]
weaviate-client = "^4.6.3"
unstructured = { extras = ["all-docs"], version = "^0.14.3" }
pdfminer-six = "^20231228"
pillow-heif = "^0.16.0"
pdf2image = "^1.17.0"
poppler-utils = "^0.1.0"
pytesseract = "^0.3.10"
pikepdf = "^9.0.0"
langchain = "^0.2.1"
langchain-community = "^0.2.1"
langchain-weaviate = "^0.0.2"
langchainhub = "^0.1.17"

[tool.poetry.scripts]
geminiplayground = "geminiplayground.cli:run"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
ruff = "^0.4.4"
mypy = "^1.10.0"

[tool.black]
line-length = 88


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
